Arista/requirements.txt
icecream==2.1.3
python-decouple==3.8
requests==2.31.0